+++ src/commands/economy-earn/challengeSubcommands/index.js
@@ -1,3 +1,4 @@
 const buttons = require("./buttons");
+const isCommand = false
 
-module.exports = { buttons };
+module.exports = { buttons, isCommand };
