+++ src/commands/helpful/commandSubcommands/index.js
@@ -1,5 +1,6 @@
 const list = require("./list");
 const disable = require("./disable");
 const enable = require("./enable");
+const isCommand = false
 
-module.exports = { list, enable, disable };
+module.exports = { list, enable, disable, isCommand };
