--- src/events/ready/01registerCommands.js
@@ -1,57 +1,0 @@
-const areCommandsDifferent = require("../../utils/areCommandsDifferent");
-const getApplicationCommands = require("../../utils/getApplicationCommands");
-const getLocalCommands = require("../../utils/getLocalCommands");
-
-module.exports = async (client) => {
-    try {
-        const localCommands = getLocalCommands();
-        const applicationCommands = await getApplicationCommands(
-            client //,
-            // testServer (this will make the commands server side again.)
-        );
-
-        for (const localCommand of localCommands) {
-            const { name, description, options } = localCommand;
-
-            const existingCommand = await applicationCommands.cache.find(
-                (cmd) => cmd.name === name
-            );
-
-            if (existingCommand) {
-                if (localCommand.deleted) {
-                    await applicationCommands.delete(existingCommand.id);
-                    console.log(`"${name}" was deleted.`);
-                    continue;
-                }
-
-                if (areCommandsDifferent(existingCommand, localCommand)) {
-                    await applicationCommands.edit(existingCommand.id, {
-                        description,
-                        options,
-                    });
-
-                    console.log(`Edited command "${name}"`);
-                }
-            } else {
-                if (localCommand.deleted) {
-                    console.log(
-                        `Skipping registering command "${name}" as if it's set to delete`
-                    );
-                    continue;
-                }
-
-                await applicationCommands.create({
-                    name,
-                    description,
-                    options,
-                });
-
-                console.log(`Registered "${name}"`);
-            }
-        }
-    } catch (error) {
-        console.log(
-            `Bruh an error (from file 01registercommands.js): ${error}`
-        );
-    }
-};
