--- src/events/ready/consoleLog.js
@@ -1,4 +1,0 @@
-
-module.exports = (client) => {
-  console.log(`${client.user.tag} is online.`);
-};
