--- src/utils/getApplicationCommands.js
@@ -1,13 +1,0 @@
-module.exports = async (client, guildId) => {
-    let applicationCommands;
-
-    if (guildId) {
-        const guild = await client.guilds.fetch(guildId);
-        applicationCommands = guild.commands;
-    } else {
-        applicationCommands = await client.application.commands;
-    }
-
-    await applicationCommands.fetch();
-    return applicationCommands;
-};
