--- src/utils/getLocalCommands.js
@@ -1,28 +1,0 @@
-/* eslint-disable no-undef */
-const path = require("path");
-const getAllFiles = require("./getAllFiles");
-
-module.exports = (exceptions = []) => {
-    let localCommands = [];
-
-    const commandCategories = getAllFiles(
-        path.join(__dirname, "..", "commands"),
-        true
-    );
-
-    for (const commandCategory of commandCategories) {
-        const commandFiles = getAllFiles(commandCategory);
-
-        for (const commandFile of commandFiles) {
-            const commandObject = require(commandFile);
-
-            if (exceptions.includes(commandObject.name)) {
-                continue;
-            }
-
-            localCommands.push(commandObject);
-        }
-    }
-
-    return localCommands;
-};
